allow files > 2GB
Hi,
I'm not entirely sure this change at itself is enough to allow files to get larger than 2GB: maybe AVI files have an internal limit or so.

In that case, maybe you want to extend the error codes a bit. Currently it returns -1 for all errors. I suggest to add e.g. -2 to indicate that the AVI limit has been reached.